fix includes and formatting
