+++ domain/include/cstone/halos/halos.hpp
@@ -55,9 +55,9 @@ namespace detail
 
 //! @brief check that only owned particles in [particleStart_:particleEnd_] are sent out as halos
 static void checkIndices(const SendList& sendList,
-                  [[maybe_unused]] LocalIndex start,
-                  [[maybe_unused]] LocalIndex end,
-                  [[maybe_unused]] LocalIndex bufferSize)
+                         [[maybe_unused]] LocalIndex start,
+                         [[maybe_unused]] LocalIndex end,
+                         [[maybe_unused]] LocalIndex bufferSize)
 {
     for (const auto& manifest : sendList)
     {
