+++ main/src/propagator/std_hydro_grackle.hpp
@@ -27,13 +27,10 @@
  * @brief A Propagator class for standard SPH including Grackle chemistry and cooling
  *
  * @author Sebastian Keller <sebastian.f.keller@gmail.com>
- * @author Jose A. Escartin <ja.escartin@gmail.com>
  * @author Noah Kubli <noah.kubli@uzh.ch>
  */
 
 #pragma once
-
-#include <variant>
 
 #include "cstone/fields/field_get.hpp"
 #include "cstone/util/value_list.hpp"
