+++ main/src/propagator/ve_hydro_bdt.hpp
@@ -41,7 +41,7 @@
 #include "sph/sph.hpp"
 #include "sph/ts_rungs.hpp"
 
-#include "propagator.hpp"
+#include "ipropagator.hpp"
 #include "gravity_wrapper.hpp"
 
 namespace sphexa
@@ -401,8 +401,7 @@ public:
                                  d.outputFieldIndices.begin();
                     transferToHost(d, first, last, {d.fieldNames[fidx]});
                     std::visit([writer, c = column, key = namesDone[i]](auto field)
-                               { writer->writeField(key, field->data(), c); },
-                               fieldPointers[fidx]);
+                               { writer->writeField(key, field->data(), c); }, fieldPointers[fidx]);
                     indicesDone.erase(indicesDone.begin() + i);
                     namesDone.erase(namesDone.begin() + i);
                 }
